+++ backend/danswer/server/query_and_chat/models.py
@@ -82,11 +82,6 @@ Currently the different branches are generated by changing the search query
 """
 
 
-class TestStandardAnswerRequest(BaseModel):
-    message: str
-    slack_bot_categories: list[str]
-
-
 class CreateChatMessageRequest(ChunkContext):
     """Before creating messages, be sure to create a chat_session and get an id"""
