+++ backend/ee/danswer/server/query_and_chat/chat_backend.py
@@ -11,17 +11,22 @@ from danswer.chat.models import DanswerAnswerPiece
 from danswer.chat.models import QADocsResponse
 from danswer.chat.models import StreamingError
 from danswer.chat.process_message import stream_chat_message_objects
+from danswer.danswerbot.slack.handlers.handle_standard_answers import (
+    oneoff_standard_answers,
+)
 from danswer.db.chat import get_or_create_root_message
 from danswer.db.engine import get_session
 from danswer.db.models import User
 from danswer.search.models import OptionalSearchSetting
 from danswer.search.models import RetrievalDetails
+from danswer.server.manage.models import StandardAnswer
 from danswer.server.query_and_chat.models import ChatMessageDetail
 from danswer.server.query_and_chat.models import CreateChatMessageRequest
 from danswer.utils.logger import setup_logger
 from ee.danswer.server.query_and_chat.models import BasicCreateChatMessageRequest
 from ee.danswer.server.query_and_chat.models import ChatBasicResponse
 from ee.danswer.server.query_and_chat.models import SimpleDoc
+from ee.danswer.server.query_and_chat.models import TestStandardAnswerRequest
 
 logger = setup_logger()
 
@@ -49,6 +54,27 @@ def remove_answer_citations(answer: str) -> str:
     pattern = r"\s*\[\[\d+\]\]\(http[s]?://[^\s]+\)"
 
     return re.sub(pattern, "", answer)
+
+
+@router.post("/test-standard-answer")
+def test_standard_answer(
+    request: TestStandardAnswerRequest,
+    db_session: Session = Depends(get_session),
+    _: User | None = Depends(current_user),
+) -> dict[str, list[StandardAnswer]] | dict[str, str]:
+    try:
+        standard_answers = oneoff_standard_answers(
+            message=request.message,
+            slack_bot_categories=request.slack_bot_categories,
+            db_session=db_session,
+        )
+        return (
+            {"standard answers": standard_answers}
+            if standard_answers
+            else {"message": "No standard answer found"}
+        )
+    except Exception as e:
+        raise HTTPException(status_code=500, detail=str(e))
 
 
 @router.post("/send-message-simple-api")
